Setup for packaging. Updated models and managers for better consistency.
