+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include scripts/*
+include static/*
