+++ mothball/db/managers/base.py
@@ -1,4 +1,4 @@
-    import abc
+import abc
 import time
 import logging
