+++ setup.py
@@ -5,7 +5,7 @@ setup(
     version = '0.1',
     packages = find_packages(),
     include_package_data = True,
-    data_files=[('etc', 'static/mothball.config'), ('bin', 'mothball.py')],
+    data_files=['static/mothball.config', 'mothball.py'],
     install_requires = [ 'sqlalchemy',
                          'boto3',
                          'pyyaml'
