+++ static/mothball.config
@@ -8,7 +8,7 @@ Database:
 
   username: mothball
 
-  password: mothballP@ssw0rd!
+  password: mothballP4ssw0rd
 
   # Host/Port only need to be defined if you are not using RDS and would rather back up to on an on prem or nonRDS db
   host:
@@ -16,7 +16,7 @@ Database:
 
 # AWS account information so that Mothball can access the api
 AWS:
-  region_name: us-east-1
+  region: us-east-1
 
   access_key: XXXXXXXXXXXXXXXXXXXX
